refactor: lint contracts

+++ src/test/RockPaperScissors.t.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.15;
 
-import "forge-std/Test.sol";
-import "forge-std/console2.sol";
-import {RockPaperScissors} from "../RockPaperScissors.sol";
+import 'forge-std/Test.sol';
+import 'forge-std/console2.sol';
+import {RockPaperScissors} from '../RockPaperScissors.sol';
 
 /*
 contract PlayerContract {
@@ -20,7 +20,6 @@ contract PlayerContract {
 */
 
 contract RockPaperScissorsTest is Test {
-
     event Game(
         address player1,
         address player2,
@@ -70,8 +69,8 @@ contract RockPaperScissorsTest is Test {
         moveAlice = uint256(Move.Scissor);
         moveBob = uint256(Move.Paper);
 
-        saltAlice = "zdqzdqdoqhj5gy3j5gj";
-        saltBob = "654qzdqd65qzdqzd6qz5d1";
+        saltAlice = 'zdqzdqdoqhj5gy3j5gj';
+        saltBob = '654qzdqd65qzdqzd6qz5d1';
 
         hashAlice = getHash(moveAlice, saltAlice);
         hashBob = getHash(moveBob, saltBob);
@@ -97,8 +96,8 @@ contract RockPaperScissorsTest is Test {
         assertEq(rps.nbrPlayer(), 0);
         assertEq(rps.player1(), address(0));
         assertEq(rps.player2(), address(0));
-        assertEq(uint(rps.movePlayer1()), 0);
-        assertEq(uint(rps.movePlayer2()), 0);
+        assertEq(uint256(rps.movePlayer1()), 0);
+        assertEq(uint256(rps.movePlayer2()), 0);
     }
 
     function testTiedGame() public {
@@ -110,8 +109,8 @@ contract RockPaperScissorsTest is Test {
         moveAlice = uint256(Move.Scissor);
         moveBob = uint256(Move.Scissor);
 
-        saltAlice = "zdqzdqdoqhj5gy3j5gj";
-        saltBob = "654qzdqd65qzdqzd6qz5d1";
+        saltAlice = 'zdqzdqdoqhj5gy3j5gj';
+        saltBob = '654qzdqd65qzdqzd6qz5d1';
 
         hashAlice = getHash(moveAlice, saltAlice);
         hashBob = getHash(moveBob, saltBob);
@@ -137,8 +136,8 @@ contract RockPaperScissorsTest is Test {
         assertEq(rps.nbrPlayer(), 0);
         assertEq(rps.player1(), address(0));
         assertEq(rps.player2(), address(0));
-        assertEq(uint(rps.movePlayer1()), 0);
-        assertEq(uint(rps.movePlayer2()), 0);
+        assertEq(uint256(rps.movePlayer1()), 0);
+        assertEq(uint256(rps.movePlayer2()), 0);
     }
 
     // Test of the register function
@@ -152,7 +151,7 @@ contract RockPaperScissorsTest is Test {
         rps.register{value: 3 ether}();
         vm.prank(charlie);
 
-        vm.expectRevert("FULL_GAME");
+        vm.expectRevert('FULL_GAME');
         rps.register{value: 3 ether}();
     }
 
@@ -160,7 +159,7 @@ contract RockPaperScissorsTest is Test {
         vm.startPrank(alice);
         rps.register{value: 3 ether}();
 
-        vm.expectRevert("INVALID_PLAYER (only another address)");
+        vm.expectRevert('INVALID_PLAYER (only another address)');
         rps.register{value: 3 ether}();
         vm.stopPrank();
     }
@@ -174,7 +173,7 @@ contract RockPaperScissorsTest is Test {
         vm.expectRevert(
             abi.encodeWithSelector(RockPaperScissors.InsufficientBet.selector, initialBet)
         );
-    
+
         rps.register{value: 1 ether}();
     }
 
@@ -194,7 +193,7 @@ contract RockPaperScissorsTest is Test {
         registerTwoPlayer();
         vm.prank(alice);
 
-        vm.expectRevert("ILLEGAL_MOVE");
+        vm.expectRevert('ILLEGAL_MOVE');
         rps.play(0);
     }
 
@@ -202,8 +201,8 @@ contract RockPaperScissorsTest is Test {
         vm.startPrank(alice);
         rps.register{value: 3 ether}();
 
-        vm.expectRevert("NO_FULL_GAME");
-        rps.play("hello you");
+        vm.expectRevert('NO_FULL_GAME');
+        rps.play('hello you');
 
         vm.stopPrank();
     }
@@ -212,23 +211,23 @@ contract RockPaperScissorsTest is Test {
         registerTwoPlayer();
 
         vm.prank(alice);
-        rps.play("f61656511551561615165");
+        rps.play('f61656511551561615165');
         vm.startPrank(bob);
-        rps.play("9843169848651164991");
+        rps.play('9843169848651164991');
 
-        vm.expectRevert("END_PLAY_PHRASE");
-        rps.play("656546465ae4a65e64ea645e");
+        vm.expectRevert('END_PLAY_PHRASE');
+        rps.play('656546465ae4a65e64ea645e');
         vm.stopPrank();
     }
 
-    // Test of the revealMove function 
+    // Test of the revealMove function
 
     function testIllegalMove() public {
         registerTwoPlayer();
 
         vm.prank(alice);
-        vm.expectRevert("ILLEGAL_REVEAL_MOVE");
-        rps.revealMove(5, "qzdqzdqzzq");
+        vm.expectRevert('ILLEGAL_REVEAL_MOVE');
+        rps.revealMove(5, 'qzdqzdqzzq');
     }
 
     function testInvalidReveal() public {
@@ -236,17 +235,17 @@ contract RockPaperScissorsTest is Test {
         playTwoPlayer();
 
         vm.prank(alice);
-        vm.expectRevert("INVALID_REVEAL");
-        rps.revealMove(uint256(Move.Scissor), "hello");
+        vm.expectRevert('INVALID_REVEAL');
+        rps.revealMove(uint256(Move.Scissor), 'hello');
     }
 
     function testWaitPlayPhase() public {
         registerTwoPlayer();
         vm.startPrank(alice);
-        rps.play("65654645ae4a65e64ea645e");
-        
-        vm.expectRevert("WAIT_REVEAL_PHASE");
-        rps.revealMove(2, "dqzdd");
+        rps.play('65654645ae4a65e64ea645e');
+
+        vm.expectRevert('WAIT_REVEAL_PHASE');
+        rps.revealMove(2, 'dqzdd');
         vm.stopPrank();
     }
 
@@ -256,7 +255,7 @@ contract RockPaperScissorsTest is Test {
         vm.deal(address(rps), intialBalance);
         registerTwoPlayer();
 
-        vm.expectRevert("GAME_LIVE");
+        vm.expectRevert('GAME_LIVE');
         vm.prank(address(0));
         rps.withdrawOwner(10 ether);
     }
@@ -264,7 +263,7 @@ contract RockPaperScissorsTest is Test {
     function testWithdrawOwnerNotOwner() public {
         vm.deal(address(rps), intialBalance);
 
-        vm.expectRevert("ONLY_OWNER");
+        vm.expectRevert('ONLY_OWNER');
         vm.prank(alice);
         rps.withdrawOwner(10 ether);
     }
@@ -278,7 +277,7 @@ contract RockPaperScissorsTest is Test {
     function testWIthdrawOwnerAvid() public {
         vm.deal(address(rps), intialBalance);
         vm.prank(address(0));
-        vm.expectRevert("INVALID_TRANSFER (withdrawOwner)");
+        vm.expectRevert('INVALID_TRANSFER (withdrawOwner)');
         rps.withdrawOwner(30 ether);
     }
 
@@ -288,7 +287,7 @@ contract RockPaperScissorsTest is Test {
         registerTwoPlayer();
 
         vm.startPrank(alice);
-        rps.play("65654645ae4a65e64ea645e");
+        rps.play('65654645ae4a65e64ea645e');
 
         vm.warp(block.timestamp + 20 minutes);
         rps.withdrawEmergency();
@@ -300,7 +299,7 @@ contract RockPaperScissorsTest is Test {
     }
 
     function withdrawEmergencyLive() public {
-        vm.expectRevert("GAME_NOT_LIVE");
+        vm.expectRevert('GAME_NOT_LIVE');
         vm.prank(alice);
         rps.withdrawEmergency();
     }
@@ -318,8 +317,8 @@ contract RockPaperScissorsTest is Test {
         moveAlice = uint256(Move.Scissor);
         moveBob = uint256(Move.Paper);
 
-        saltAlice = "zdqzdqdoqhj5gy3j5gj";
-        saltBob = "654qzdqd65qzdqzd6qz5d1";
+        saltAlice = 'zdqzdqdoqhj5gy3j5gj';
+        saltBob = '654qzdqd65qzdqzd6qz5d1';
 
         hashAlice = getHash(moveAlice, saltAlice);
         hashBob = getHash(moveBob, saltBob);
